Quick, DoubleSelection, Selection, Bogo Sorts Animation. Mild Refactoring (coloring columns "coloredTimeLine"). Recolor the array after finishing with green color "end". Move Class has new boolean attribute "pivot" for quick sort so the pivot can have unique color. some "Ctrl + Alt + L".

+++ src/main/java/Main/Implementation/Sorting/Sorting.java
@@ -9,13 +9,13 @@ import java.util.List;
 import java.util.Queue;
 
 public class Sorting implements Runnable {
-    List<Double> toSort;
-    Comparator<Double> comparator;
-    SortAttributes<Double> sort;
+    List<Integer> toSort;
+    Comparator<Integer> comparator;
+    SortAttributes<Integer> sort;
     Queue<Move> moves;
 
-    public Sorting(List<Double> toSort, Comparator<Double> comparator,
-                   SortAttributes<Double> sort, Queue<Move> moves) {
+    public Sorting(List<Integer> toSort, Comparator<Integer> comparator,
+                   SortAttributes<Integer> sort, Queue<Move> moves) {
         this.toSort = toSort;
         this.comparator = comparator;
         this.sort = sort;
