+++ src/main/java/Main/Implementation/Sorting/SortingStrategies/InsertionSort.java
@@ -14,7 +14,7 @@ public class InsertionSort<T> extends SortAttributes<T> {
             int j = i - 1;
             while (j >= 0 && comparator.compare(toSort.get(j), current) > 0) {
                 if (animate) {
-                    buffer = new Move(j + 1, j, (Double) current, (Double) toSort.get(j), true);
+                    buffer = new Move(j + 1, j, (Integer) current, (Integer) toSort.get(j), true);
                     moves.add(buffer);
                 }
                 toSort.set(j + 1, toSort.get(j));
