+++ src/main/java/Main/Implementation/Sorting/SortingStrategies/OptimizedBubbleSort.java
@@ -17,7 +17,7 @@ public class OptimizedBubbleSort<T> extends SortAttributes<T> {
             for (int j = 0; j < i; j++) {
                 int cmp = comparator.compare(toSort.get(j), toSort.get(j + 1));
                 if (animate)
-                    buffer = new Move(j + 1, j, (Double) toSort.get(j + 1), (Double) toSort.get(j), false);
+                    buffer = new Move(j + 1, j, (Integer) toSort.get(j + 1), (Integer) toSort.get(j), false);
 
                 if (cmp > 0) {
                     if (animate)
